+++ src/modules/admin/widgets/grids/columns/FileThumbnailColumn.php
@@ -3,27 +3,14 @@
 namespace davidhirtz\yii2\media\modules\admin\widgets\grids\columns;
 
 use davidhirtz\yii2\media\models\File;
-use davidhirtz\yii2\skeleton\helpers\Html;
-use yii\grid\DataColumn;
+use davidhirtz\yii2\skeleton\modules\admin\widgets\grids\columns\LinkDataColumn;
 
-class FileThumbnailColumn extends DataColumn
+class FileThumbnailColumn extends LinkDataColumn
 {
-    /**
-     * @var int|null the width of the column
-     */
-    public ?int $columnWidth = 150;
-
-    /**
-     * @var callable|null a callback function that returns the route for the count link
-     */
-    public mixed $route = null;
+    public $headerOptions = ['style' => 'width:150px'];
 
     public function init(): void
     {
-        if ($this->columnWidth) {
-            Html::addCssStyle($this->headerOptions, ['width' => "{$this->columnWidth}px"]);
-        }
-
         if (!is_callable($this->content)) {
             $this->content = $this->renderThumbnail(...);
         }
@@ -36,25 +23,6 @@ class FileThumbnailColumn extends DataColumn
      */
     protected function renderThumbnail(File $model, int $key, int $index): string
     {
-        $content = $this->renderThumbnailContent($model);
-        $route = is_callable($this->route) ? call_user_func($this->route, $model) : $this->route;
-
-        if ($route) {
-            $content = Html::a($content, $route);
-        }
-
-        return $content;
-    }
-
-    protected function renderThumbnailContent(File $file): string
-    {
-        if (!$file->hasPreview()) {
-            return '';
-        }
-
-        return Html::tag('div', '', [
-            'style' => 'background-image:url(' . ($file->getTransformationUrl('admin') ?: $file->getUrl()) . ');',
-            'class' => 'thumb',
-        ]);
+        return Thumbnail::widget(['file' => $model]);
     }
 }